fastAPI-deploy/requirements.txt
alembic==1.8.1 
anyio==3.6.2
asyncpg==0.27.0
click==8.1.3
databases==0.6.2
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.5
opencv-python==4.6.0.66
orjson==3.8.3
Pillow==9.3.0
psycopg2==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.22.0
tritonclient[all]
typing_extensions==4.4.0
ujson==5.6.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4